requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LangChain and LangGraph
langchain>=0.3.0
langchain-aws>=0.2.0
langgraph>=0.2.0

# AWS
boto3>=1.34.0
botocore>=1.34.0

# FastAPI and server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Document processing
pypdf>=4.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Vector store (optional)
faiss-cpu>=1.7.4

# Validation and retry
tenacity>=8.2.0
jsonschema>=4.21.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Utilities
click>=8.1.0
rich>=13.7.0
tqdm>=4.66.0

# Original dependencies (kept for compatibility)
torch>=2.0.0
transformers>=4.36.0
tokenizers>=0.15.0
einops
addict
easydict
